fix: dva don't work after build and enabled dynamicImport
