+++ packages/umi-build-dev/template/history.js.tpl
@@ -1,0 +1,2 @@
+// create history
+window.g_history = {{{ history }}};
