+++ packages/umi-plugin-dva/template/DvaContainer.js
@@ -1,26 +1,9 @@
 import { Component } from 'react';
-import dva from 'dva';
-import createLoading from 'dva-loading';
-
-const runtimeDva = window.g_plugins.mergeConfig('dva');
-let app = dva({
-  history: window.g_history,
-  <%= ExtendDvaConfig %>
-  ...(runtimeDva.config || {}),
-});
-<%= EnhanceApp %>
-window.g_app = app;
-app.use(createLoading());
-(runtimeDva.plugins || []).forEach(plugin => {
-  app.use(plugin);
-});
-<%= RegisterPlugins %>
-<%= RegisterModels %>
 
 class DvaContainer extends Component {
   render() {
-    app.router(() => this.props.children);
-    return app.start()();
+    window.g_app.router(() => this.props.children);
+    return window.g_app.start()();
   }
 }
